--- plugin-build/src/main/resources/META-INF/gradle-plugins/io.sentry.android.gradle.properties
@@ -1,1 +1,0 @@
-implementation-class=io.sentry.android.gradle.SentryPlugin
