+++ plugin-build/src/test/resources/testFixtures/appTestProject/app/build.gradle
@@ -1,0 +1,6 @@
+plugins {
+    id "com.android.application"
+    id "io.sentry.android.gradle"
+}
+
+apply plugin: 'com.android.application'
