+++ plugin-build/src/test/resources/testFixtures/appTestProject/app/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.example" />
