+++ plugin-build/src/test/resources/testFixtures/appTestProject/build.gradle
@@ -1,0 +1,2 @@
+// Note that this file will be overridden
+// by tests to populate the plugin classpath.
