+++ plugin-build/src/test/resources/testFixtures/appTestProject/gradle.properties
@@ -1,0 +1,2 @@
+org.gradle.daemon=false
+org.gradle.jvmargs=-Xmx512m -XX\:MaxMetaspaceSize\=512m
