+++ plugin-build/src/test/resources/testFixtures/appTestProject/sentry.properties
@@ -1,0 +1,3 @@
+defaults.org=example
+defaults.project=example
+auth.token=deadd00d
